chore : git 충돌 해결
eslint 잘못 지워서 발생한 충돌 해결
